refactor: extract core extension navigation policy hooks
