+++ addons/smart_construction_core/core_extension.py
@@ -63,6 +63,33 @@ from odoo.addons.smart_construction_core.core_extension_intents import (
 from odoo.addons.smart_construction_core.core_extension_system_init import (
     apply_system_init_profile_overrides,
 )
+from odoo.addons.smart_construction_core.core_extension_navigation_policy import (
+    smart_core_business_config_admin_group_xmlids,
+    smart_core_business_config_form_settings_refs,
+    smart_core_business_config_approval_policy_refs,
+    smart_core_native_config_root_menu_xmlid,
+    smart_core_native_config_delivery_excluded_menu_xmlids,
+    smart_core_lowcode_system_config_menu_xmlids,
+    smart_core_lowcode_config_recovery_parent_menu_xmlids,
+    smart_core_business_root_menu_xmlid,
+    smart_core_relation_entry_policy,
+    smart_core_model_specific_form_contract_policy,
+    smart_core_form_field_aliases,
+    smart_core_menu_delivery_token_policy,
+    smart_core_business_nav_group_display_order,
+    smart_core_product_policy_catalog_source,
+    smart_core_product_policy_catalog_base_keys,
+    smart_core_default_product_policy_specs,
+    smart_core_product_policy_catalog_label,
+    smart_core_platform_legacy_ownership_module,
+    smart_core_resolve_release_actor_role_codes,
+    smart_core_resolve_usage_actor_role_codes,
+    smart_core_default_release_snapshot_role_code,
+    smart_core_industry_extension_module_names,
+    smart_core_app_shell_contract,
+    smart_core_scene_entry_orchestrator_specs,
+    smart_core_user_data_acceptance_nav_contract,
+)
 from odoo.addons.smart_construction_core.core_extension_workspace_facts import (
     _build_enterprise_enablement_contract,
     _build_home_block_contract_rows,
@@ -2342,679 +2369,3 @@ def smart_core_build_settlement_slice_service(env):
     )
 
     return SettlementSliceService(env)
-
-
-def smart_core_business_config_admin_group_xmlids(env):
-    del env
-    return [
-        "smart_construction_core.group_sc_cap_business_config_admin",
-        "smart_construction_core.group_sc_role_business_admin",
-    ]
-
-
-def smart_core_business_config_form_settings_refs(env):
-    del env
-    return {
-        "action_xmlid": "smart_construction_core.action_ui_form_field_policy_business_config",
-        "menu_xmlid": "smart_construction_core.menu_ui_form_field_policy_business_config",
-    }
-
-
-def smart_core_business_config_approval_policy_refs(env):
-    del env
-    return {
-        "action_xmlid": "smart_construction_core.action_sc_approval_policy",
-        "menu_xmlid": "smart_construction_core.menu_sc_approval_policy",
-    }
-
-
-def smart_core_native_config_root_menu_xmlid(env):
-    del env
-    return "smart_construction_core.menu_sc_business_config_center"
-
-
-def smart_core_native_config_delivery_excluded_menu_xmlids(env):
-    del env
-    return [
-        "smart_construction_core.menu_project_quota_root",
-        "smart_construction_core.menu_project_quota_center",
-        "smart_construction_core.menu_project_quota_subitem",
-        "smart_construction_core.menu_project_quota_tree",
-        "smart_construction_core.menu_sc_dictionary_root",
-        "smart_construction_core.menu_sc_dictionary_all",
-        "smart_construction_core.menu_sc_dictionary_discipline",
-        "smart_construction_core.menu_sc_dictionary_chapter",
-        "smart_construction_core.menu_sc_dictionary_quota_item",
-        "smart_construction_core.menu_sc_dictionary_sub_item",
-        "smart_construction_core.menu_quota_import_wizard",
-    ]
-
-
-def smart_core_lowcode_system_config_menu_xmlids(env):
-    del env
-    return [
-        "smart_construction_core.menu_sc_business_config_center",
-        "smart_construction_core.menu_sc_business_base_config_group",
-        "smart_construction_core.menu_sc_lowcode_system_config_group",
-        "smart_construction_core.menu_sc_business_category",
-        "smart_construction_core.menu_sc_dictionary",
-        "smart_construction_core.menu_sc_organization_department",
-        "smart_construction_core.menu_sc_approval_scope",
-        "smart_construction_core.menu_sc_approval_policy",
-        "smart_construction_core.menu_sc_project_stage_requirement_items",
-        "smart_construction_core.menu_sc_project_cost_code",
-        "smart_construction_core.menu_sc_company_document_archive",
-        "smart_construction_core.menu_sc_business_config_workbench",
-        "smart_construction_core.menu_ui_menu_config_policy_business_config",
-        "smart_construction_core.menu_ui_form_field_policy_business_config",
-        "smart_construction_core.menu_ui_form_custom_field_wizard_business_config",
-    ]
-
-
-def smart_core_lowcode_config_recovery_parent_menu_xmlids(env):
-    del env
-    return [
-        "smart_construction_core.menu_sc_business_config_center",
-    ]
-
-
-def smart_core_business_root_menu_xmlid(env):
-    del env
-    return "smart_construction_core.menu_sc_root"
-
-
-def smart_core_relation_entry_policy(env, payload):
-    payload = payload if isinstance(payload, dict) else {}
-    model = _sc_text(payload.get("model"))
-    field_name = _sc_text(payload.get("field_name"))
-    relation = _sc_text(payload.get("relation"))
-    context = payload.get("context") if isinstance(payload.get("context"), dict) else {}
-    record = payload.get("record") if isinstance(payload.get("record"), dict) else {}
-    if relation == "account.tax" and field_name == "tax_id" and model in {"construction.contract", "sc.general.contract"}:
-        default_vals = {
-            "type_tax_use": "none",
-            "amount_type": "percent",
-            "price_include": False,
-        }
-        try:
-            company = env.company
-            helper = env["construction.contract"].sudo().with_company(company)
-            group = helper._sc_contract_tax_group(company)
-            country = company.account_fiscal_country_id or company.partner_id.country_id or env.ref(
-                "base.cn",
-                raise_if_not_found=False,
-            )
-            default_vals.update({
-                "company_id": company.id,
-                "tax_group_id": group.id,
-            })
-            if country:
-                default_vals["country_id"] = country.id
-        except Exception:
-            _logger.debug("Unable to resolve contract tax quick-create defaults.", exc_info=True)
-        return {
-            "has_page": False,
-            "can_open": False,
-            "can_create": True,
-            "quick_create": True,
-            "display_field": "name",
-            "order": "amount asc, id asc",
-            "default_vals": default_vals,
-            "domain": [
-                ["type_tax_use", "=", "none"],
-                ["amount_type", "=", "percent"],
-                ["price_include", "=", False],
-            ],
-            "ui_labels": {
-                "quick_create": "新增税率...",
-                "create_and_edit": "新增税率...",
-                "missing_name": "请输入税率百分比，例如：3%、9%、13%",
-                "quick_create_prompt": "请输入税率百分比，例如：3%、9%、13%",
-                "quick_create_failed": "新增税率失败，请输入 0 到 100 之间的百分比",
-                "inline_create": "新增税率“%s”",
-                "inline_create_failed": "创建税率失败，请检查百分比格式",
-            },
-        }
-    if relation == "res.partner" and model == "sc.self.funding.registration" and field_name == "partner_id":
-        category_code = _sc_text(context.get("current_business_category_code") or context.get("default_business_category_code"))
-        if category_code == "finance.self_funding.refund":
-            contractor_ids = []
-            try:
-                summaries = env["sc.company.contractor.responsibility.summary"].sudo().search([
-                    ("partner_id", "!=", False),
-                    ("self_funding_balance", ">", 0.01),
-                ])
-                partners = summaries.mapped("partner_id").filtered(
-                    lambda partner: partner.supplier_rank > 0 or partner.customer_rank > 0
-                )
-                contractor_ids = sorted(set(partners.ids))
-            except Exception:
-                contractor_ids = []
-            return {
-                "has_page": False,
-                "can_create": False,
-                "display_field": "display_name",
-                "order": "name asc, id asc",
-                "domain": [["id", "in", contractor_ids or [0]]],
-                "ui_labels": {
-                    "search_more": "搜索承包人...",
-                    "dialog_title": "承包人：搜索更多",
-                    "search_placeholder": "输入承包人名称搜索",
-                    "create_and_edit": "",
-                    "quick_create": "",
-                },
-            }
-    if relation == "res.partner" and model == "sc.expense.claim" and field_name == "partner_id":
-        category_code = _sc_text(context.get("current_business_category_code") or context.get("default_business_category_code"))
-        if category_code == "finance.deduction.bill":
-            partner_ids = []
-            try:
-                summaries = env["sc.company.contractor.responsibility.summary"].sudo().search([("partner_id", "!=", False)])
-                partners = summaries.mapped("partner_id").filtered(
-                    lambda partner: partner.supplier_rank > 0 or partner.customer_rank > 0
-                )
-                partner_ids = sorted(set(partners.ids))
-            except Exception:
-                partner_ids = []
-            return {
-                "has_page": False,
-                "can_create": False,
-                "display_field": "display_name",
-                "order": "name asc, id asc",
-                "domain": [["id", "in", partner_ids or [0]]],
-                "ui_labels": {
-                    "search_more": "搜索责任方...",
-                    "dialog_title": "责任方：搜索更多",
-                    "search_placeholder": "输入责任方名称搜索",
-                    "create_and_edit": "",
-                    "quick_create": "",
-                },
-            }
-    if relation == "construction.contract" and model == "construction.contract" and field_name == "original_contract_id":
-        category_code = _sc_text(context.get("current_business_category_code") or context.get("default_business_category_code"))
-        contract_type = _sc_text(context.get("default_type") or context.get("type") or record.get("type"))
-        domain = []
-        if category_code == "contract.income.supplement":
-            domain = [["type", "=", "out"]]
-        elif category_code == "contract.expense.supplement":
-            domain = [["type", "=", "in"]]
-        elif contract_type in {"out", "in"}:
-            domain = [["type", "=", contract_type]]
-        if not domain:
-            return None
-        return {
-            "has_page": False,
-            "can_open": False,
-            "can_create": False,
-            "create_mode": "disabled",
-            "display_field": "display_name",
-            "order": "id desc",
-            "domain": domain,
-            "ui_labels": {
-                "search_more": "搜索原合同...",
-                "dialog_title": "原合同：搜索更多",
-                "search_placeholder": "输入原合同名称、编号、项目或往来单位搜索",
-                "create_and_edit": "",
-                "quick_create": "",
-            },
-        }
-    return None
-
-
-def smart_core_model_specific_form_contract_policy(env, payload):
-    del env
-    payload = payload if isinstance(payload, dict) else {}
-    model = _sc_text(payload.get("model"))
-    fields_map = payload.get("fields") if isinstance(payload.get("fields"), dict) else {}
-    if model == "sc.general.contract" and "tax_id" in fields_map and "tax_rate" in fields_map:
-        return {"remove_fields": ["tax_rate"]}
-    return None
-
-
-def smart_core_form_field_aliases(env, payload):
-    del env
-    payload = payload if isinstance(payload, dict) else {}
-    model = _sc_text(payload.get("model"))
-    source = payload.get("source_contract") if isinstance(payload.get("source_contract"), dict) else {}
-    fields_map = source.get("fields") if isinstance(source.get("fields"), dict) else {}
-    if model == "sc.general.contract" and "tax_id" in fields_map:
-        return {"tax_rate": "tax_id"}
-    return None
-
-
-def smart_core_menu_delivery_token_policy(env):
-    del env
-    business_tokens = [
-        "定额库",
-        "成本科目",
-        "阶段资料要求",
-        "项目阶段资料要求",
-    ]
-    user_tokens = [
-        "智能施工 2.0",
-        "项目管理",
-        "合同管理",
-        "合同中心",
-        "项目合同",
-        "收入合同",
-        "支出合同",
-        "一般合同",
-        "施工管理",
-        "计划管理",
-        "计划汇报",
-        "成本管理",
-        "成本报表",
-        "智慧大屏",
-        "大屏",
-        "驾驶舱",
-        "财税报表",
-        "物资与分包",
-        "物资管理",
-        "材料管理",
-        "材料",
-        "采购",
-        "采购订单",
-        "劳务",
-        "机械",
-        "设备",
-        "专业分包",
-        "分包",
-        "考勤",
-        "看板中心",
-        "财务账款",
-        "结算中心",
-        "执行结构",
-        "项目立项",
-        "项目台账",
-        "项目驾驶舱",
-        "投标",
-        "报名",
-        "开标",
-        "中标",
-        "保证金",
-        "自筹",
-        "付款还",
-        "资金借还",
-        "借款",
-        "还款",
-        "费用中心",
-        "费用",
-        "报销",
-        "收支",
-        "统计表",
-        "经营情况表",
-        "收入",
-        "公司财务支出",
-        "项目资金",
-        "承包人",
-        "项目款",
-        "公司款",
-        "发票税务",
-        "发票",
-        "开票",
-        "税务",
-        "付款",
-        "实付",
-        "付款登记",
-        "支付",
-        "往来单位付款",
-        "扣款",
-        "实缴",
-        "资金账户",
-        "账户间资金往来",
-        "收款",
-        "到款",
-        "资金日报",
-        "工程资料",
-        "客户",
-        "供应商",
-        "人事行政",
-        "请假",
-        "休假",
-        "印章",
-        "资料证照",
-        "证照",
-        "借阅",
-        "业务配置",
-        "配置中心",
-        "组织架构",
-        "历史用户",
-        "历史用户权限",
-        "用户信息",
-        "用户信息与权限",
-        "历史角色",
-        "项目授权范围",
-        "系统配置",
-        "用户验收",
-        "直营项目系统菜单",
-    ]
-    admin_tokens = [
-        "执行与生产",
-        "成本与资金",
-    ]
-    return {
-        "always_hidden_technical_tokens": ["项目管理（后台）"],
-        "business_config_tokens": business_tokens,
-        "user_allowed_path_tokens": user_tokens,
-        "admin_allowed_path_tokens": [*user_tokens, *admin_tokens],
-        "hide_exact_labels": [
-            "快速创建项目",
-            "项目列表（演示）",
-        ],
-        "rename_labels": {
-            "项目台账（试点）": "项目台账",
-            "开票申请": "销项开票申请",
-            "开票登记": "销项发票登记",
-            "进项上报": "进项税额上报",
-        },
-    }
-
-
-def smart_core_business_nav_group_display_order(env):
-    del env
-    return {
-        "基础资料": 5,
-        "项目中心": 10,
-        "投标管理类单据": 20,
-        "合同中心": 30,
-        "施工管理": 40,
-        "物资与分包": 50,
-        "财务中心": 60,
-        "人事行政": 70,
-        "资料证照": 80,
-        "配置中心": 990,
-        "配置": 990,
-        "系统配置": 990,
-    }
-
-
-def smart_core_product_policy_catalog_source(env, source_env=None):
-    del env, source_env
-    return {
-        "module": "smart_construction_core",
-        "xmlid_prefix": "smart_construction_core.",
-        "root_label": "智慧施工管理平台",
-    }
-
-
-def smart_core_product_policy_catalog_base_keys(env):
-    del env
-    return ["construction"]
-
-
-def smart_core_default_product_policy_specs(env):
-    del env
-    return [
-        ("construction.standard", "施工管理标准版"),
-        ("construction.preview", "施工管理预览版"),
-    ]
-
-
-def smart_core_product_policy_catalog_label(env, identity):
-    del env
-    edition_key = str((identity or {}).get("edition_key") or "").strip()
-    return "施工管理预览版" if edition_key == "preview" else "施工管理标准版"
-
-
-def smart_core_platform_legacy_ownership_module(env):
-    del env
-    return "smart_construction_core"
-
-
-def smart_core_resolve_release_actor_role_codes(env, user):
-    del env
-    if not user:
-        return []
-    roles = set()
-    try:
-        group_xmlids = set(user.groups_id.get_external_id().values())
-    except Exception:
-        group_xmlids = set()
-    prefix = "smart_construction_core.group_sc_role_"
-    for xmlid in group_xmlids:
-        text = str(xmlid or "").strip()
-        if text.startswith(prefix):
-            roles.add(text[len(prefix):])
-    try:
-        if user.has_group("smart_construction_core.group_sc_cap_project_read") or user.has_group(
-            "smart_construction_core.group_sc_cap_project_manager"
-        ):
-            roles.add("pm")
-        if user.has_group("smart_construction_core.group_sc_super_admin") or user.has_group(
-            "smart_construction_core.group_sc_business_full"
-        ):
-            roles.add("executive")
-    except Exception:
-        _logger.debug("Unable to resolve release actor role codes from groups.", exc_info=True)
-    return sorted(roles)
-
-
-def smart_core_resolve_usage_actor_role_codes(env, user):
-    return smart_core_resolve_release_actor_role_codes(env, user)
-
-
-def smart_core_default_release_snapshot_role_code(env):
-    del env
-    return "pm"
-
-
-def smart_core_industry_extension_module_names(env):
-    del env
-    return [
-        "smart_construction_core",
-        "smart_construction_scene",
-        "smart_construction_portal",
-    ]
-
-
-def smart_core_app_shell_contract(env):
-    del env
-    return {
-        "taxonomy": {
-            "dashboard": {"label": "经营驾驶舱", "category": "management", "sequence": 10, "primary_scene": "dashboard.company"},
-            "projects": {"label": "项目管理", "category": "business", "sequence": 20, "primary_scene": "projects.list"},
-            "contracts": {"label": "合同管理", "category": "business", "sequence": 30, "primary_scene": "contracts.workspace"},
-            "cost": {"label": "成本管理", "category": "business", "sequence": 40, "primary_scene": "cost.cost_compare"},
-            "finance": {"label": "资金财务", "category": "business", "sequence": 50, "primary_scene": "finance.payment_requests"},
-            "payments": {"label": "收支办理", "category": "business", "sequence": 60, "primary_scene": "payments.approval"},
-            "operation": {"label": "运营管理", "category": "business", "sequence": 80, "primary_scene": "operation.overview"},
-            "portfolio": {"label": "项目组合", "category": "management", "sequence": 90, "primary_scene": "portfolio.center"},
-            "risk": {"label": "风险管理", "category": "governance", "sequence": 100, "primary_scene": "risk.center"},
-            "quality": {"label": "质量管理", "category": "business", "sequence": 110, "primary_scene": "quality.center"},
-            "safety": {"label": "安全管理", "category": "business", "sequence": 120, "primary_scene": "safety.center"},
-            "resource": {"label": "资源管理", "category": "business", "sequence": 130, "primary_scene": "resource.center"},
-            "task": {"label": "任务协同", "category": "productivity", "sequence": 140, "primary_scene": "task.center"},
-        },
-        "aliases": {
-            "project": "projects",
-            "contract": "contracts",
-            "payment": "payments",
-        },
-    }
-
-
-def smart_core_scene_entry_orchestrator_specs(env):
-    del env
-    common_project_summary = ("project_code", "manager_name", "stage_name")
-    return {
-        "ProjectDashboardSceneOrchestrator": {
-            "scene_key": "project.management",
-            "scene_label": "项目驾驶舱",
-            "state_fallback_text": "后端未提供项目驾驶舱状态摘要",
-            "title_empty": "项目驾驶舱",
-            "title_template": "项目驾驶舱：{name}",
-            "title_record_name_fallback": "项目",
-            "suggested_action_key": "load_dashboard_progress",
-            "suggested_action_reason_code": "PROJECT_DASHBOARD_READY",
-            "block_fetch_intent": "project.dashboard.block.fetch",
-            "block_alias_map": {"risk": "risks"},
-            "first_action_block_keys": ["progress"],
-            "entry_summary_keys": (
-                "project_code",
-                "manager_name",
-                "partner_name",
-                "stage_name",
-                "lifecycle_state",
-                "milestone",
-                "health_state",
-                "progress_percent",
-                "cost_total",
-                "payment_total",
-                "status",
-            ),
-            "entry_blocks": (
-                ("progress", "项目进度", "deferred"),
-                ("risks", "风险提醒", "deferred"),
-                ("next_actions", "下一步动作", "deferred"),
-            ),
-        },
-        "ProjectExecutionSceneOrchestrator": {
-            "scene_key": "project.execution",
-            "scene_label": "项目执行",
-            "state_fallback_text": "后端未提供项目执行状态摘要",
-            "title_empty": "项目执行",
-            "title_template": "项目执行：{name}",
-            "title_record_name_fallback": "项目",
-            "suggested_action_key": "load_execution_next_actions",
-            "suggested_action_reason_code": "PROJECT_EXECUTION_READY",
-            "block_fetch_intent": "project.execution.block.fetch",
-            "first_action_block_keys": ["next_actions", "execution_tasks"],
-            "entry_summary_keys": common_project_summary + ("date_start", "date_end"),
-            "entry_blocks": (
-                ("execution_tasks", "执行任务", "deferred"),
-                ("readiness_precheck", "上线前检查", "deferred"),
-                ("next_actions", "下一步动作", "deferred"),
-            ),
-        },
-        "ProjectPlanBootstrapSceneOrchestrator": {
-            "scene_key": "project.plan_bootstrap",
-            "scene_label": "计划准备",
-            "state_fallback_text": "后端未提供计划准备状态摘要",
-            "title_empty": "计划编排",
-            "title_template": "计划准备：{name}",
-            "title_record_name_fallback": "项目",
-            "suggested_action_key": "load_plan_next_actions",
-            "suggested_action_reason_code": "PROJECT_PLAN_BOOTSTRAP_READY",
-            "block_fetch_intent": "project.plan_bootstrap.block.fetch",
-            "first_action_block_keys": ["next_actions", "plan_summary_detail"],
-            "entry_summary_keys": common_project_summary + ("date_start", "date_end"),
-            "entry_blocks": (
-                ("plan_summary_detail", "计划摘要", "deferred"),
-                ("plan_tasks", "计划任务", "deferred"),
-                ("next_actions", "计划下一步", "deferred"),
-            ),
-        },
-        "PaymentSliceContractOrchestrator": {
-            "scene_key": "payment",
-            "scene_label": "支付记录",
-            "state_fallback_text": "后端未提供支付记录状态摘要",
-            "title_empty": "支付记录",
-            "title_template": "支付记录：{name}",
-            "title_record_name_fallback": "项目",
-            "suggested_action_key": "load_payment_entry",
-            "suggested_action_reason_code": "PAYMENT_SLICE_PREPARED_READY",
-            "block_fetch_intent": "payment.block.fetch",
-            "first_action_block_keys": ["payment_entry", "payment_list"],
-            "entry_summary_keys": common_project_summary + ("payment_record_count", "payment_total_amount"),
-            "entry_blocks": (
-                ("payment_entry", "支付录入", "deferred"),
-                ("payment_list", "支付记录", "deferred"),
-                ("payment_summary", "支付汇总", "deferred"),
-                ("next_actions", "支付下一步", "deferred"),
-            ),
-        },
-        "SettlementSliceContractOrchestrator": {
-            "scene_key": "settlement",
-            "scene_label": "结算结果",
-            "state_fallback_text": "后端未提供结算结果状态摘要",
-            "title_empty": "结算结果",
-            "title_template": "结算结果：{name}",
-            "title_record_name_fallback": "项目",
-            "suggested_action_key": "load_settlement_summary",
-            "suggested_action_reason_code": "SETTLEMENT_SLICE_PREPARED_READY",
-            "block_fetch_intent": "settlement.block.fetch",
-            "entry_summary_keys": common_project_summary + ("total_cost", "total_payment", "delta"),
-            "entry_blocks": (
-                ("settlement_summary", "结算结果", "deferred"),
-                ("next_actions", "结算下一步", "deferred"),
-            ),
-        },
-        "CostTrackingContractOrchestrator": {
-            "scene_key": "cost.tracking",
-            "scene_label": "成本记录",
-            "state_fallback_text": "后端未提供成本记录状态摘要",
-            "title_empty": "成本记录",
-            "title_template": "成本记录：{name}",
-            "title_record_name_fallback": "项目",
-            "suggested_action_key": "load_cost_entry",
-            "suggested_action_reason_code": "COST_SLICE_PREPARED_READY",
-            "block_fetch_intent": "cost.tracking.block.fetch",
-            "first_action_block_keys": ["cost_entry", "cost_list"],
-            "entry_summary_keys": common_project_summary + ("cost_record_count", "cost_total_amount"),
-            "entry_blocks": (
-                ("cost_entry", "成本录入", "deferred"),
-                ("cost_list", "成本记录", "deferred"),
-                ("cost_summary", "成本汇总", "deferred"),
-                ("next_actions", "成本下一步", "deferred"),
-            ),
-        },
-    }
-
-
-def smart_core_user_data_acceptance_nav_contract(env):
-    del env
-    return {
-        "formal_group_child_labels": ["客户", "供应商"],
-        "formal_group_labels": ["配置中心"],
-        "old_acceptance_group_labels": ["用户核对菜单", "旧业务数据核对"],
-        "direct_acceptance_group_labels": ["直营项目数据核对", "直营项目系统菜单"],
-        "joint_acceptance_group_labels": ["联营项目数据核对", "联营项目系统菜单"],
-        "direct_acceptance_child_tokens": ["直营"],
-        "joint_acceptance_child_tokens": ["联营"],
-        "acceptance_root_labels": ["用户验收", "直营项目数据核对"],
-        "acceptance_root_group_label": "用户数据验收",
-        "direct_acceptance_group_label": "直营项目数据核对",
-        "joint_acceptance_group_label": "联营项目数据核对",
-        "source_menu_group_labels_to_hide": ["用户核对菜单"],
-        "acceptance_surface_menu_ids": [727, 729, 735, 770],
-        "acceptance_surface_action_ids": [899],
-        "acceptance_surface_tokens": [
-            "用户验收",
-            "用户数据验收",
-            "用户核对菜单",
-            "menu_scbsly_acceptance_",
-            "menu_scbs55_user_acceptance_",
-            "menu_scbsly_direct_project_acceptance_root",
-            "menu_sc_legacy_engineering_progress_receipt",
-        ],
-        "old_acceptance_menu_xmlids": [
-            "smart_construction_core.menu_scbs55_user_acceptance_445_工程进度收款",
-        ],
-        "joint_acceptance_menu_xmlids": [
-            "smart_construction_core.menu_scbsly_joint_acceptance_self_funding_advance_income",
-            "smart_construction_core.menu_scbsly_joint_acceptance_self_funding_advance_refund",
-            "smart_construction_core.menu_scbsly_joint_acceptance_supplier_contract",
-            "smart_construction_core.menu_scbsly_joint_acceptance_labor_contract",
-            "smart_construction_core.menu_scbsly_joint_acceptance_rental_contract",
-        ],
-        "contract_product_menu_xmlids": [
-            "smart_construction_core.menu_sc_contract_income",
-            "smart_construction_core.menu_sc_project_income_contract",
-            "smart_construction_core.menu_sc_income_contract_execution",
-            "smart_construction_core.menu_sc_contract_event",
-            "smart_construction_core.menu_sc_general_contract",
-            "smart_construction_core.menu_sc_contract_expense",
-            "smart_construction_core.menu_sc_expense_contract_execution",
-        ],
-        "settlement_product_menu_xmlids": [
-            "smart_construction_core.menu_sc_settlement_order",
-            "smart_construction_core.menu_sc_settlement_adjustment",
-            "smart_construction_core.menu_sc_income_contract_settlement",
-            "smart_construction_core.menu_sc_expense_contract_settlement",
-            "smart_construction_core.menu_sc_material_settlement",
-            "smart_construction_core.menu_sc_labor_settlement",
-            "smart_construction_core.menu_sc_equipment_settlement",
-            "smart_construction_core.menu_sc_material_rental_settlement",
-            "smart_construction_core.menu_sc_subcontract_settlement",
-        ],
-    }
